refactor: 이벤트 결과 페이지 데이터 구조 개선
